+++ plugin-locationcomponent/src/main/java/com/mapbox/maps/plugin/locationcomponent/animators/Evaluators.kt
@@ -8,10 +8,17 @@ internal object Evaluators {
    * Type evaluator for Point data
    */
   val POINT = TypeEvaluator<Point> { fraction, startValue, endValue ->
-    Point.fromLngLat(
-      startValue.longitude() + fraction * (endValue.longitude() - startValue.longitude()),
-      startValue.latitude() + fraction * (endValue.latitude() - startValue.latitude())
-    )
+    val longitude = startValue.longitude() + fraction * (endValue.longitude() - startValue.longitude())
+    val latitude = startValue.latitude() + fraction * (endValue.latitude() - startValue.latitude())
+    if (startValue.hasAltitude() && endValue.hasAltitude()) {
+      Point.fromLngLat(
+        longitude,
+        latitude,
+        startValue.altitude() + fraction * (endValue.altitude() - startValue.altitude()),
+      )
+    } else {
+      Point.fromLngLat(longitude, latitude)
+    }
   }
 
   /**